athena/modules/programming/module_programming_llm/pyproject.toml
[tool.poetry]
name = "module_programming_llm"
version = "0.1.0"
description = "Programming assessment LLM module."
authors = ["Felix Dietrich <felixtj.dietrich@tum.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.11.*"
# if you have local changes in the common Athena module, use the line below. Otherwise, please use a VCS stable version. Also, a version with tag = "" is possible.
# athena = { path = "../athena", develop = true }
athena = { path = "../../../athena", develop = true }
#athena = { git = "https://github.com/ls1intum/Athena.git", rev = "2da2d33", subdirectory = "athena"}
openai = "1.40.0"
langchain = "0.3.2"
langchain-community="0.3.1"
langchain-openai = "0.2.1"
gitpython = "^3.1.41"
tiktoken = "0.7.0"
promptlayer = "^0.1.85"
python-dotenv = "^1.0.0"
nltk = "^3.8.1"
replicate = "^0.11.0"
chromadb = "0.5.13"
streamlit= "1.25.0"
html2text = "2020.1.16"
google-api-core = "2.11.1"
google-api-python-client = "2.95.0"
google-auth = "2.22.0"
google-auth-httplib2 = "0.1.0"
googleapis-common-protos = "1.59.1"
faiss-cpu = {version = "^1.7.0", platform = "posix"}
faiss-gpu = {version = "^1.7.1", platform = "posix"}
bs4 = "0.0.2"

[tool.poetry.group.dev.dependencies]

pydantic = "2.7.4"
prospector = "^1.10.2"

[tool.poetry.scripts]
module = "athena:run_module"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
